+++ modes/AudioApps/BuntyAudioApp.hpp
@@ -454,7 +454,7 @@ protected:
     float verbVsDelayLevel{0}, delayToVerbLevel{0}, filterBankDelayXFade{0};
     float delayMorph{0.5f}, delayBlend{0.f};
 
-    OnePoleSmoother<kN_Params> smoother{150.f, kSampleRate};
+    OnePoleSmoother<kN_Params> smoother{150.f, (float)kSampleRate};
 
     daisysp::PitchShifter pitchshifter_;
     float pitchshifter_mix_{0.5f};
