+++ modes/AudioApps/VerbFXAudioApp.hpp
@@ -346,7 +346,7 @@ protected:
     float verbVsDelayLevel{0}, delayToVerbLevel{0}, filterBankDelayXFade{0};
     float delayMorph{0.5f}, delayBlend{0.f};
 
-    OnePoleSmoother<kN_Params> smoother{150.f, kSampleRate};
+    OnePoleSmoother<kN_Params> smoother{150.f, (float)kSampleRate};
     
     // maxiDynamicsLite limiter;
